renamed loader to createList
