+++ createList/geo-ip.go
@@ -1,4 +1,4 @@
-package loader
+package createList
 
 //Reads in CSV file and creates a node list
