+++ createList/geo-ip_test.go
@@ -1,20 +1,20 @@
-package loader_test
+package createList_test
 
 import (
 	"os"
 	"reflect"
 	"testing"
 
-	"github.com/m-lab/annotation-service/loader"
+	"github.com/m-lab/annotation-service/createList"
 )
 
 
 //tests correct parsing of createList
 func TestCreateList(t *testing.T) {
 	r, _ := os.Open("testdata/sample.csv")
-	list, _ := loader.CreateList(r)
-	var listComp = []loader.Node{
-		loader.Node{
+	list, _ := createList.CreateList(r)
+	var listComp = []createList.Node{
+		createList.Node{
 			"1.0.1.0",
 			"1.0.3.255",
 			16777472,
@@ -22,7 +22,7 @@ func TestCreateList(t *testing.T) {
 			"CN",
 			"China",
 		},
-		loader.Node{
+		createList.Node{
 			"1.0.4.0",
 			"1.0.7.255",
 			16778240,
@@ -30,7 +30,7 @@ func TestCreateList(t *testing.T) {
 			"AU",
 			"Australia",
 		},
-		loader.Node{
+		createList.Node{
 			"1.0.8.0",
 			"1.0.15.255",
 			16779264,
@@ -38,7 +38,7 @@ func TestCreateList(t *testing.T) {
 			"CN",
 			"China",
 		},
-		loader.Node{
+		createList.Node{
 			"1.0.16.0",
 			"1.0.31.255",
 			16781312,
